+++ pysui/sui/sui_clients/sync_client.py
@@ -18,7 +18,8 @@ from typing import Any, Union
 from json import JSONDecodeError
 import httpx
 from pysui.sui.sui_clients.common import _ClientMixin, PreExecutionResult, SuiRpcResult
-from pysui.sui.sui_types.scalars import ObjectID, SuiInteger, SuiTxBytes, SuiString
+from pysui.sui.sui_crypto import MultiSig, SuiPublicKey
+from pysui.sui.sui_types.scalars import ObjectID, SuiInteger, SuiSignature, SuiTxBytes, SuiString
 from pysui.sui.sui_types.address import SuiAddress
 from pysui.sui.sui_types.collections import SuiArray, SuiMap
 from pysui.sui.sui_txresults.single_tx import FaucetGasRequest
@@ -33,7 +34,6 @@ from pysui.sui.sui_builders.get_builders import (
     GetObject,
     GetPackage,
     GetEvents,
-    GetTxs,
 )
 from pysui.sui.sui_builders.exec_builders import (
     DryRunTransaction,
@@ -109,9 +109,28 @@ class SuiClient(_ClientMixin):
                 result = SuiRpcResult(
                     True, None, PreExecutionResult(builder.authority, builder.handle_return(result["result"]))
                 )
-                #    SuiTxBytes(result["result"]["txBytes"])))
             return result
         return SuiRpcResult(False, "execute_no_sign is used only with transaction types")
+
+    def execute_with_multisig(
+        self, builder: SuiBaseBuilder, multi_sig: MultiSig, pub_keys: list[SuiPublicKey]
+    ) -> Union[SuiRpcResult, Exception]:
+        """execute_with_multisig Executes a transaction, signing with MultiSig."""
+        if builder.txn_required:
+            result = self.execute_no_sign(builder)
+            if result.is_ok():
+                tx_bytes = result.result_data.tx_bytes
+                new_sig = multi_sig.sign(tx_bytes, pub_keys)
+                if isinstance(new_sig, SuiSignature):
+                    exec_tx = ExecuteTransaction(
+                        tx_bytes=tx_bytes,
+                        signatures=SuiArray([new_sig]),
+                        request_type=SuiRequestType.WAITFORLOCALEXECUTION,
+                    )
+                    return self.execute(exec_tx)
+                return SuiRpcResult(False, f"Signing error code {new_sig}")
+            return result
+        return self._execute(builder)
 
     def sign_and_submit(
         self, signer: SuiAddress, tx_bytes: SuiTxBytes, additional_signatures: SuiArray[SuiAddress] = None
@@ -227,15 +246,12 @@ class SuiClient(_ClientMixin):
         """
         for_address = for_address or self.config.active_address
         try:
-            print(self.config.faucet_url)
-            print({"FixedAmountRequest": {"recipient": f"{for_address}"}})
             result = self._client.post(
                 self.config.faucet_url,
                 headers=GetObjectsOwnedByAddress(for_address).header,
                 json={"FixedAmountRequest": {"recipient": f"{for_address}"}},
             ).json()
             if result["error"] is None:
-                print(result)
                 return SuiRpcResult(True, None, FaucetGasRequest.from_dict(result))
             return SuiRpcResult(False, result["error"])
         except JSONDecodeError as jexc:
@@ -310,24 +326,6 @@ class SuiClient(_ClientMixin):
         inargs: dict = locals().copy()
         inargs.pop("self")
         return self.execute(GetEvents(**inargs))
-
-    def get_txns(self, *, query: SuiMap, cursor: str, limit: int, descending_order: bool) -> SuiRpcResult:
-        """get_txns `sui_getTransactions` API.
-
-        :param query: The transaction query type map.
-        :type query: SuiMap
-        :param cursor: _description_
-        :type cursor: str
-        :param limit: _description_
-        :type limit: int
-        :param descending_order: _description_
-        :type descending_order: bool
-        :return: API call result
-        :rtype: SuiRpcResult
-        """
-        inargs: dict = locals().copy()
-        inargs.pop("self")
-        return self.execute(GetTxs(**inargs))
 
     def pay_txn(
         self,